#55 ImageUpload Local class added for testing

+++ src/main/java/ch/saunah/saunahbackend/util/ImageUploadUtil.java
@@ -1,5 +1,8 @@
 package ch.saunah.saunahbackend.util;
 
+import com.amazonaws.AmazonClientException;
+import com.amazonaws.AmazonServiceException;
+import com.amazonaws.SdkClientException;
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
@@ -10,7 +13,6 @@ import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.S3Object;
 import com.amazonaws.services.s3.model.S3ObjectInputStream;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -20,8 +22,7 @@ import java.nio.file.Paths;
 /**
  * This class is used as a helper class to save and read images from a directory.
  */
-@Component
-public class ImageUploadUtil {
+public class ImageUploadUtil implements ImageUpload {
 
     @Value("${saunah.object.storage.bucket.id}")
     private String bucket;
@@ -45,6 +46,12 @@ public class ImageUploadUtil {
         return s3client;
     }
 
+    private String getObjectPath(String directory, String fileName){
+        Path uploadPath = Paths.get(directory);
+        Path filePath = uploadPath.resolve(fileName);
+        return filePath.toString();
+    }
+
     /**
      * Saves the image to the specified directory.
      *
@@ -53,15 +60,20 @@ public class ImageUploadUtil {
      * @param multipartFile the image object
      * @throws IOException throws when Path is not valid
      */
-    public void saveImage(String directory, String fileName, MultipartFile multipartFile) {
-        AmazonS3 client = getObjectStorageClient();
-        if (!client.doesBucketExist(bucket)) {
-            client.createBucket(bucket);
+    public void saveImage(String directory, String fileName, MultipartFile multipartFile) throws IOException {
+        try {
+            AmazonS3 client = getObjectStorageClient();
+            if (!client.doesBucketExist(bucket)) {
+                client.createBucket(bucket);
+            }
+            ObjectMetadata data = new ObjectMetadata();
+            data.setContentType(multipartFile.getContentType());
+            data.setContentLength(multipartFile.getSize());
+            client.putObject(bucket, getObjectPath(directory, fileName), multipartFile.getInputStream(), data);
         }
-        ObjectMetadata data = new ObjectMetadata();
-        data.setContentType(multipartFile.getContentType());
-        data.setContentLength(multipartFile.getSize());
-        client.putObject(bucket, getObjectPath(directory, fileName), multipartFile.getInputStream(), data);
+        catch (AmazonClientException e){
+            throw new IOException(e.getMessage());
+        }
     }
 
     /**
@@ -73,10 +85,17 @@ public class ImageUploadUtil {
      * @throws IOException throws when the path is not valid
      */
     public byte[] getImage(String directory, String fileName) throws IOException {
-        AmazonS3 client = getObjectStorageClient();
-        S3Object s3object = client.getObject(bucket, getObjectPath(directory, fileName));
-        S3ObjectInputStream inputStream = s3object.getObjectContent();
-        return inputStream.readAllBytes();
+        try {
+            AmazonS3 client = getObjectStorageClient();
+            if (!client.doesBucketExist(bucket)) {
+                throw new IOException("Bucket does not exist and no image can be found!");
+            }
+            S3Object s3object = client.getObject(bucket, getObjectPath(directory, fileName));
+            S3ObjectInputStream inputStream = s3object.getObjectContent();
+            return inputStream.readAllBytes();
+        } catch (AmazonClientException e) {
+            throw new IOException(e.getMessage());
+        }
     }
 
     /**
@@ -85,14 +104,15 @@ public class ImageUploadUtil {
      * @param directory the directory where the file is stored
      * @param fileName the filename of the image
      */
-    public void removeImage(String directory, String fileName){
+    public void removeImage(String directory, String fileName) throws IOException {
+        try {
         AmazonS3 client = getObjectStorageClient();
+        if (!client.doesBucketExist(bucket)) {
+            throw new IOException("Bucket does not exist and no image can be found!");
+        }
         client.deleteObject(bucket, getObjectPath(directory, fileName));
-    }
-
-    private String getObjectPath(String directory, String fileName){
-        Path uploadPath = Paths.get(directory);
-        Path filePath = uploadPath.resolve(fileName);
-        return filePath.toString();
+        } catch (AmazonClientException e) {
+            throw new IOException(e.getMessage());
+        }
     }
 }
